Coloquei a função de conversão dentro do arquivo
