+++ funcoes.h
@@ -1,2 +1,3 @@
 float conversor_volume(float valor, int unidade_inicial, int unidade_final);
 float Converter_Temperatura(float valor, int opcao_entrada, int opcao_final);
+float conversor_armazenamento(float valor, int unidade_inicial, int unidade_final);
